requirements.txt
Flask==3.0.0
Flask-MySQLdb==1.0.1
python-dotenv==1.0.0
